2nd project
A linear regression model capable of predicting the salaries of employees using their years of experience on the job. The dataset used contains 30 samples. The dataset was divided into TRAINING(80%) AND TESTING (20%) datasets. The model was trained on the TRAINING DATA and its performance was evaluated by testing with the TESTING DATASET.